Removed bug of inconsistent parsing into words, testing still needed.
Originally, the code was getting the trust, origin, and sigs by parsing annotated markup.
The problem is that on rare occasions, the parsing of the annotated markup did not yield the same
sequence of words as the parsing of the original text.
So now, I only parse the original text, and I get the trust, origin, and sigs information
by writing it into the wikitrust_sigs table.
Testing still needed.