Added dbsocket to MySQL opening call
These changes fix a bug in which WikiTrust failed to build on my system.
One errors included that dbsocket was missing from every place that a
MySQL connection was opened. I was able to build the software by adding
variable, set to None, to every place this was called.

My understanding is that sockets are optional in new version of MySQL
but can be used instead of the standard port based method.

I have not yet tested this patch with a database and I am currently
using WikiTrust to parse XML dumps without a MySQL database present.